delete(status changing) and file upload work in purchaserOrder , invoice

+++ src/main/java/com/invoicemanagement/service/InvoiceService.java
@@ -9,7 +9,7 @@ public interface InvoiceService {
 
 	public Invoice create(Map<String,Object> invoice);
 	public List<Invoice> getAll();
-	public void delete(long id); 
+	public void delete(int id); 
 	public Invoice update(Invoice invoice, long id);
 	public Invoice getById(int id);
 }
